Move type aliases
