+++ py-polars/polars/convert.py
@@ -11,9 +11,8 @@ from polars.internals import DataFrame, Series
 from polars.utils.decorators import deprecate_nonkeyword_arguments, deprecated_alias
 
 if TYPE_CHECKING:
-    from polars.datatypes import SchemaDefinition, SchemaDict
     from polars.dependencies import numpy as np
-    from polars.type_aliases import Orientation
+    from polars.type_aliases import Orientation, SchemaDefinition, SchemaDict
 
 
 @deprecated_alias(columns="schema")
