+++ py-polars/polars/expr/string.py
@@ -13,9 +13,8 @@ from polars.datatypes import (
 )
 
 if TYPE_CHECKING:
-    from polars.datatypes import PolarsDataType, PolarsTemporalType
     from polars.expr.expr import Expr
-    from polars.type_aliases import TransferEncoding
+    from polars.type_aliases import PolarsDataType, PolarsTemporalType, TransferEncoding
 
 
 class ExprStringNameSpace:
