+++ py-polars/polars/functions/eager.py
@@ -26,11 +26,15 @@ if TYPE_CHECKING:
     from datetime import date
 
     from polars.dataframe.frame import DataFrame
-    from polars.datatypes import PolarsDataType
     from polars.expr.expr import Expr
     from polars.lazyframe.frame import LazyFrame
     from polars.series.series import Series
-    from polars.type_aliases import ClosedInterval, ConcatMethod, TimeUnit
+    from polars.type_aliases import (
+        ClosedInterval,
+        ConcatMethod,
+        PolarsDataType,
+        TimeUnit,
+    )
 
     if sys.version_info >= (3, 8):
         from typing import Literal
