+++ py-polars/polars/lazyframe/groupby.py
@@ -8,9 +8,8 @@ from polars.utils.decorators import deprecated_alias
 
 if TYPE_CHECKING:
     from polars.dataframe.frame import DataFrame
-    from polars.datatypes import SchemaDict
     from polars.polars import PyLazyGroupBy
-    from polars.type_aliases import IntoExpr, RollingInterpolationMethod
+    from polars.type_aliases import IntoExpr, RollingInterpolationMethod, SchemaDict
 
 LDF = TypeVar("LDF", bound="pli.LazyFrame")
